requirements.txt
streamlit
sentence-transformers
faiss-cpu
pdfplumber
python-docx
pandas
python-dotenv
google-genai
